+++ models/sale_detail_model.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel, Field
 from datetime import datetime
+from pydantic import BaseModel, Field
+from decimal import Decimal
 
 
 class Product(BaseModel):
@@ -14,6 +15,6 @@ class SaleDetail(BaseModel):
   product: Product = Field(...)
   units: int = Field(...)
   unity_price: int = Field(...)
-  total_price: float = Field(...)
+  total_price: Decimal = Field(..., decimal_places=2)
   created_at: datetime = None
   updated_at: datetime = None
