+++ models/sale_model.py
@@ -1,6 +1,7 @@
+from datetime import datetime
 from enum import Enum
 from pydantic import BaseModel, Field
-from datetime import datetime
+from decimal import Decimal
 
 
 class PayWith(str, Enum):
@@ -12,7 +13,7 @@ class PayWith(str, Enum):
 
 class Sale(BaseModel):
   id: str = Field(..., alias="_id")
-  total_amount: float = Field(...)
+  total_amount: Decimal = Field(..., decimal_places=2)
   pay_type: PayWith = Field(...)
   customer: str = Field(...)
   created_at: datetime = None
